pyproject.toml
[tool.poetry]
name = "aily-crm"
version = "0.1.0"
description = ""
authors = ["JG <gaojunqi@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aily-sdk = "^0.30.0"
arrow = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
